requirements.txt
ddddocr~=1.0.6
requests~=2.31.0
pycryptodome~=3.19.0
